dnsdist: Move rules definitions to a separate header
It makes it easier to test them.
